add tests and improvements
